requirements.txt
asgiref==3.2.3
boto3==1.10.10
botocore==1.13.10
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-storages==1.9.1
django-uuidfield==0.5.0
djangorestframework==3.11.0
djangorestframework-gis==0.14
docutils==0.15.2
fcm-django==0.3.2
idna==2.9
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.5
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.2
Pillow==7.0.0
pyfcm==1.4.7
python-dateutil==2.8.0
pytz==2019.3
razorpay==1.2.0
requests==2.23.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
tz==0.2.2
uritemplate==3.0.0
urllib3==1.25.8
